+++ GRDB/Core/Database+Schema.swift
@@ -132,6 +132,7 @@ extension Database {
     /// The number of columns in the table named `tableName`.
     ///
     /// - throws: A DatabaseError if table does not exist.
+    @available(*, deprecated, message: "Use db.columns(in: tableName).count instead")
     public func columnCount(in tableName: String) throws -> Int {
         return try columns(in: tableName).count
     }